requirements.txt
customtkinter~=5.2.2
pillow~=10.2.0
pytube~=15.0.0